[Refactor]apply-compound-components-NotiBar #781
